started combining components
